Remove English language module and associated ARIA files from the project
