+++ frontend/src/community/common/assets/aria-labels/en/attendance.json
