+++ frontend/src/community/common/assets/aria-labels/en/english.ts
@@ -1,0 +1,14 @@
+import attendanceAria from "~community/common/assets/aria-labels/en/attendance.json";
+import commonAria from "~community/common/assets/aria-labels/en/common.json";
+import leaveAria from "~community/common/assets/aria-labels/en/leave.json";
+import onboardingAria from "~community/common/assets/aria-labels/en/onboarding.json";
+import peopleAria from "~community/common/assets/aria-labels/en/people.json";
+
+export const englishAria = {
+  onboardingAria,
+  attendanceAria,
+  commonAria,
+  leaveAria,
+  peopleAria
+  // add other modules here
+};
