+++ frontend/src/community/common/assets/aria-labels/en/onboarding.json
