+++ frontend/src/community/common/assets/languages/en/configurations.json
