+++ frontend/src/community/common/assets/languages/en/english.ts
@@ -1,0 +1,16 @@
+import attendanceModule from "~community/common/assets/languages/en/attendanceModule.json";
+import common from "~community/common/assets/languages/en/common.json";
+import configurations from "~community/common/assets/languages/en/configurations.json";
+import leaveModule from "~community/common/assets/languages/en/leaveModule.json";
+import onboarding from "~community/common/assets/languages/en/onboarding.json";
+import peopleModule from "~community/common/assets/languages/en/peopleModule.json";
+
+export const english = {
+  ...common,
+  onboarding,
+  configurations,
+  peopleModule,
+  leaveModule,
+  attendanceModule
+  // add other modules here
+};
