--- frontend/src/community/common/assets/languages/english/english.ts
@@ -1,26 +1,0 @@
-import attendanceAria from "~community/common/assets/languages/english/aria/attendance.json";
-import commonAria from "~community/common/assets/languages/english/aria/common.json";
-import leaveAria from "~community/common/assets/languages/english/aria/leave.json";
-import onboardingAria from "~community/common/assets/languages/english/aria/onboarding.json";
-import peopleAria from "~community/common/assets/languages/english/aria/people.json";
-import attendanceModule from "~community/common/assets/languages/english/attendanceModule.json";
-import common from "~community/common/assets/languages/english/common.json";
-import configurations from "~community/common/assets/languages/english/configurations.json";
-import leaveModule from "~community/common/assets/languages/english/leaveModule.json";
-import onboarding from "~community/common/assets/languages/english/onboarding.json";
-import peopleModule from "~community/common/assets/languages/english/peopleModule.json";
-
-export const english = {
-  ...common,
-  onboarding,
-  configurations,
-  peopleModule,
-  leaveModule,
-  attendanceModule,
-  onboardingAria,
-  attendanceAria,
-  commonAria,
-  leaveAria,
-  peopleAria
-  // add other modules here
-};
